+++ CHANGELOG.md
@@ -6,8 +6,8 @@
   * Add error code and error message for `exceededTimeoutLimit`  
 * BraintreePayPal
   * Add PayPal edit funding instrument flow (BETA)
-    * Add `BTPayPalEditRequest` for edit FI flow
-    * Add `BTPayPalClient.tokenize(_:completion:)` method that takes in a `BTPayPalEditRequest`
+    * Add `BTPayPalVaultEditRequest` for edit FI flow
+    * Add `BTPayPalClient.edit(_:completion:)` method that takes in a `BTPayPalVaultEditRequest`
     * **Note:** This feature is currently in beta and may change or be removed in future releases
 
 ## 6.23.0 (2024-07-15)